ci(lint): migrates pluto linting tool to use docker image
Previously, Pluto was retrieved using Homebrew, however this made the CI dependent on running on a MacOS based runner.

This change migrates the linter to use the official Docker images, which allows us to use Ubuntu based runners.

One advantage is that this should make it easier to also run the tests locally, specifically now also tools like 'act' can be used to run the pipeline locally.