Add markdown linting
This adds a new GitHub Action to lint all markdown files in the project.

A configuration file limits line lengths to 100 charactes (default is 80) and excludes code blocks
and tables from this limit. It also allows the <details> and <summary> tags, since there are no
good alternatives.

Linting is disabled for anything in `/old` and `/scenarios`.

Finally vscode will suggest adding the markdownlint extension when opening the project. This will
help vscode users to get linting warnings _before_ committing.